state: split into modules and reorganize
Closes #975.
